# Choose num_threads in parallel_for based on GRAIN_SIZE