--- a/aten/src/ATen/ParallelOpenMP.h
+++ b/aten/src/ATen/ParallelOpenMP.h
@@ -27,15 +27,15 @@ inline void parallel_for(
   std::exception_ptr eptr;
   // choose number of tasks based on grain size and number of threads
   int64_t num_threads = omp_in_parallel() ? 1 : omp_get_max_threads();
-  const int64_t num_iter = end - begin;
   if (grain_size > 0) {
-    num_threads = std::min(num_threads, divup(num_iter, grain_size));
+    num_threads = std::min(num_threads, divup((end - begin), grain_size));
   }
-  const int64_t chunk_size = divup(num_iter, num_threads);
 
-#pragma omp parallel if(num_threads > 1) num_threads(num_threads)
+#pragma omp parallel num_threads(num_threads)
   {
+    int64_t num_threads = omp_get_num_threads();
     int64_t tid = omp_get_thread_num();
+    int64_t chunk_size = divup((end - begin), num_threads);
     int64_t begin_tid = begin + tid * chunk_size;
     if (begin_tid < end) {
       try {