@soumith 2019-09-27
unfortunately that didn't fix it. I am thinking out loud, but wonder if there are correctness bugs wrt this patch