@soumith 2019-09-30
the results of computations are just plain wrong, like nans and large positive / negative numbers etc., which seems like bad OpenMP side-effects