@peterbell10 2019-09-30
Okay, looking at the [OpenMP spec] I see that the `num_threads` clause is only an upper bound on the actual number of threads launched when `OMP_DYNAMIC` is used. So, I've removed all code that made that faulty assumption. I can't say whether that's the issue in the tests but it would certainly cause weird issues if the number of threads wasn't as expected.

I also see from the spec that the effect of an `if` clause is to set 
```
number of threads = 1;
```
so it shouldn't change anything compared to `num_threads(1)`.

[OpenMP spec]: https://www.openmp.org/spec-html/5.0/openmpsu35.html