@ezyang 2019-09-30
@soumith, delegating final review to you