@soumith 2019-10-03
@peterbell10 this is starting to look good. didn't see any failures on limited tests with clang + openmp, and just to be sure I started a more comprehensive test suite