@JianpingChen066 2019-10-16
@peterbell10  @jgong5 @Jianhui-Li
Hi, Peter
This check-in cause a little regression on the [DLRM](https://github.com/facebookresearch/dlrm) benchmark

Without this check-in the DLRM benchmark result is like:
numactl --physcpubind=0-27 -m 0 python dlrm_s_pytorch.py --mini-batch-size=2048 --num-batches=100 --data-generation=random --arch-mlp-bot=512-512-64 --arch-mlp-top=1024-1024-1024-1 --arch-sparse-feature-size=64 --arch-embedding-size=1000000-1000000-1000000-1000000-1000000-1000000-1000000-1000000 --num-indices-per-lookup=100 --arch-interaction-op=dot --numpy-rand-seed=727 --print-freq=100 --print-time --enable-profiling > model1_CPU_PT_28.log
Min time per iteration = 3591.85

with this check-in , the result is like:
numactl --physcpubind=0-27 -m 0 python dlrm_s_pytorch.py --mini-batch-size=2048 --num-batches=100 --data-generation=random --arch-mlp-bot=512-512-64 --arch-mlp-top=1024-1024-1024-1 --arch-sparse-feature-size=64 --arch-embedding-size=1000000-1000000-1000000-1000000-1000000-1000000-1000000-1000000 --num-indices-per-lookup=100 --arch-interaction-op=dot --numpy-rand-seed=727 --print-freq=100 --print-time --enable-profiling > model1_CPU_PT_28.log
Min time per iteration = 3649.57

The profiling data show that there are some time increased in the 
index_select , mm and addmm operations

                       This              Before
index_select    1559.31ms   1503.48ms
             mm    27.14ms       14.58ms
       addmm    22.38ms       9.79966ms

Please have a check, thanks