@peterbell10 2019-10-16
@JianpingChen066 that probably means that `grain_size` needs to be tuned differently for those operations.