@JianpingChen066 2019-10-17
The new issue is submitted at https://github.com/pytorch/pytorch/issues/28198
Thanks