@JianpingChen066 2019-10-18
It's may better to have two parameters for here, one for threshold to launch the parallel version and one for grain size to partition (if this is 0, then balance the workload among the threads)